+++ src/Http/Controllers/Dominion/StatusController.php
@@ -6,6 +6,7 @@ use OpenDominion\Calculators\Dominion\LandCalculator;
 use OpenDominion\Calculators\Dominion\MilitaryCalculator;
 use OpenDominion\Calculators\Dominion\PopulationCalculator;
 use OpenDominion\Calculators\NetworthCalculator;
+use OpenDominion\Helpers\NotificationHelper;
 use OpenDominion\Services\Dominion\ProtectionService;
 use OpenDominion\Services\Dominion\QueueService;
 
@@ -13,13 +14,20 @@ class StatusController extends AbstractDominionController
 {
     public function getStatus()
     {
+        $resultsPerPage = 25;
+        $selectedDominion = $this->getSelectedDominion();
+
+        $notifications = $selectedDominion->notifications()->paginate($resultsPerPage);
+
         return view('pages.dominion.status', [
             'dominionProtectionService' => app(ProtectionService::class),
             'landCalculator' => app(LandCalculator::class),
             'militaryCalculator' => app(MilitaryCalculator::class),
             'networthCalculator' => app(NetworthCalculator::class),
+            'notificationHelper' => app(NotificationHelper::class),
             'populationCalculator' => app(PopulationCalculator::class),
-            'queueService' => app(QueueService::class)
+            'queueService' => app(QueueService::class),
+            'notifications' => $notifications
         ]);
     }
 }